@@ -9,80 +9,68 @@ import (
 
 const JSMAXSAFE = 1342177
 
-func SRandom(seed... uint64)float64{
-	if len(seed)==0 {
+func SRandom(seed ...uint64) float64 {
+	if len(seed) == 0 {
 		return rand.Float64()
 	}
-	s1:=seed[0]
-	s1%=JSMAXSAFE
-	s1 = (s1 + 1627) * (s1 + 125383)+s1*2
-	s1%=JSMAXSAFE
-	s1+=s1 * 9311 + 149297
-	s1%=JSMAXSAFE
+	s1 := seed[0]
+	s1 %= JSMAXSAFE
+	s1 = (s1+1627)*(s1+125383) + s1*2
+	s1 %= JSMAXSAFE
+	s1 += s1*9311 + 149297
+	s1 %= JSMAXSAFE
 
-	s2:=uint64(1)
-	y:=13
-	for y>0 {
-		s2*=s1
-		s2%=JSMAXSAFE
+	s2 := uint64(1)
+	y := 13
+	for y > 0 {
+		s2 *= s1
+		s2 %= JSMAXSAFE
 		y--
 	}
 	return float64(s2) / float64(JSMAXSAFE)
 }
 
-func Rng(val1,val2 float64,seed...uint64)float64{
-	var minVal,maxVal float64
-	if val1<val2 {
+func Rng[T Number](val1, val2 T, seed ...uint64) T {
+	var minVal, maxVal T
+	if val1 < val2 {
 		minVal = val1
 		maxVal = val2
 	} else {
 		minVal = val2
 		maxVal = val1
 	}
-	return minVal+(maxVal-minVal)*SRandom(seed...)
+	return minVal + T(float64(maxVal-minVal)*SRandom(seed...))
 }
 
-func RngInt(val1,val2 int,seed...uint64)int {
-	return int(math.Floor(Rng(float64(val1),float64(val2),seed...)))
-}
-
-func Porn(seed...uint64)float64{
-	if OneIn(2.0,seed...) {
+func Porn(seed ...uint64) float64 {
+	if OneIn(2.0, seed...) {
 		return -1.0
 	} else {
 		return 1.0
 	}
 }
 
-func PornInt(seed...uint64)int{
-	if OneIn(2.0,seed...) {
+func PornInt(seed ...uint64) int {
+	if OneIn(2.0, seed...) {
 		return -1
 	} else {
 		return 1
 	}
 }
 
-func OneIn(chance float64,seed...uint64)bool {
-	return chance<=1||Rng(0,chance,seed...)<1.0
+func OneIn[T Number](chance T, seed ...uint64) bool {
+	return chance <= 1 || Rng(0, 1/float64(chance), seed...) < 1.0
 }
 
-func OneInInt(chance int,seed...uint64)bool {
-	return chance<=1||RngInt(0,chance,seed...)<1.0
-}
-
-func XInY(x float64,y float64,seed...uint64)bool {
-	return SRandom(seed...)<x/y
-}
-
-func XInYInt(x int,y int,seed...uint64)bool {
-	return SRandom(seed...)<float64(x)/float64(y)
+func XInY[T Number](x T, y T, seed ...uint64) bool {
+	return SRandom(seed...) < float64(x)/float64(y)
 }
 
 type WeightAble interface {
-	Weight()float64
+	Weight() float64
 }
 
-func WeightSelect(weightList []WeightAble,seed...uint64)(int,WeightAble) {
+func WeightSelect[T WeightAble](weightList []T, seed ...uint64) (int, T) {
 	if len(weightList) == 0 {
 		log.Panic("weight list must > 0")
 	}
@@ -100,27 +88,27 @@ func WeightSelect(weightList []WeightAble,seed...uint64)(int,WeightAble) {
 			r -= weight
 		}
 	}
-	return i,weightList[i]
+	return i, weightList[i]
 }
 
-func Shuffle(arr []interface{},seed...uint64)[]interface{}{
+func Shuffle[T any](arr []T, seed ...uint64) []T {
 	sr := true
-	if len(seed)==0 {
+	if len(seed) == 0 {
 		sr = false
 	}
-	leng:=len(arr)
-	var temp interface{}
+	leng := len(arr)
+	var temp T
 	var random int
 	seedMod := 0
-	collection:=slice.SliceConcat(arr)
-	for leng>0 {
+	collection := slice.Concat(arr)
+	for leng > 0 {
 		seedMod++
 		if sr {
-			random = int(math.Floor(SRandom(seed[0]+uint64(seedMod))*float64(leng)))
+			random = int(math.Floor(SRandom(seed[0]+uint64(seedMod)) * float64(leng)))
 		} else {
-			random = int(math.Floor(rand.Float64()*float64(leng)))
+			random = int(math.Floor(rand.Float64() * float64(leng)))
 		}
-		leng-=1
+		leng -= 1
 		temp = collection[leng]
 		collection[leng] = collection[random]
 		collection[random] = temp
@@ -128,50 +116,50 @@ func Shuffle(arr []interface{},seed...uint64)[]interface{}{
 	return collection
 }
 
-func GetOne(arr []interface{},seed...uint64)interface{}{
-	index := RngInt(0,len(arr)-1,seed...)
+func GetOne[T any](arr []T, seed ...uint64) T {
+	index := Rng(0, len(arr)-1, seed...)
 	return arr[index]
 }
 
-func PickOne(arr []interface{},seed...uint64)(interface{},[]interface{}){
-	leng:=len(arr)
-	index:=int(math.Floor(float64(leng)*SRandom(seed...)))
-	var ret interface{}
-	retOrigin:=make([]interface{},0)
-	for i:=0;i<leng;i++ {
-		if i==index {
+func PickOne[T any](arr []T, seed ...uint64) (T, []T) {
+	leng := len(arr)
+	index := int(math.Floor(float64(leng) * SRandom(seed...)))
+	var ret T
+	retOrigin := make([]T, 0)
+	for i := 0; i < leng; i++ {
+		if i == index {
 			ret = arr[i]
 			continue
 		}
 		retOrigin = append(retOrigin, arr[i])
 	}
-	return ret,retOrigin
+	return ret, retOrigin
 }
 
-func PickSome(arr []interface{},num int,seed...uint64)([]interface{},[]interface{}){
+func PickSome[T any](arr []T, num int, seed ...uint64) ([]T, []T) {
 	sr := true
-	if len(seed)==0 {
+	if len(seed) == 0 {
 		sr = false
 	}
-	leng:=len(arr)
+	leng := len(arr)
 	var index int
-	retOrigin:=make([]interface{},0)
-	ret:=make([]interface{},0)
-	spliceArr :=make([]int,0)
-	for i:=0;i<num;i++ {
+	retOrigin := make([]T, 0)
+	ret := make([]T, 0)
+	spliceArr := make([]int, 0)
+	for i := 0; i < num; i++ {
 		if sr {
-			index =int(math.Floor(float64(leng)*rand.Float64()))
+			index = int(math.Floor(float64(leng) * rand.Float64()))
 		} else {
-			index = int(math.Floor(float64(leng)*SRandom(seed[0]+uint64(i))))
+			index = int(math.Floor(float64(leng) * SRandom(seed[0]+uint64(i))))
 		}
 		spliceArr = append(spliceArr, index)
 	}
- 	for i:=0;i<leng;i++ {
- 		if slice.HasInt(spliceArr,i) {
+	for i := 0; i < leng; i++ {
+		if slice.Has(spliceArr, i) {
 			ret = append(ret, arr[i])
- 			continue
+			continue
 		}
 		retOrigin = append(retOrigin, arr[i])
 	}
-	return ret,retOrigin
+	return ret, retOrigin
 }
