@@ -5,54 +5,48 @@ import (
 	"math"
 )
 
-func PowInt(x,y int)int {
-	ret:=1
-	for y>0 {
-		ret*=x
+type Number interface {
+	uint8 | int8 | uint16 | int16 | uint32 | int32 | uint64 | int64 | int | float32 | float64
+}
+
+type Signed interface {
+	int8 | int16 | int32 | int64 | int | float32 | float64
+}
+
+type Float interface {
+	float32 | float64
+}
+
+func Pow[T Number](x, y T) T {
+	var ret T = 1
+	for y > 0 {
+		ret *= x
 		y--
 	}
 	return ret
 }
 
-func PowUInt64(x,y uint64)uint64 {
-	ret:=uint64(1)
-	for y>0 {
-		ret*=x
-		y--
-	}
-	return ret
-}
-
-func PowInt64(x,y int64)int64 {
-	ret:=int64(1)
-	for y>0 {
-		ret*=x
-		y--
-	}
-	return ret
-}
-
-func MaxInt(x,y int)int{
-	if x>y {
+func Max[T Number](x, y T) T {
+	if x > y {
 		return x
 	}
 	return y
 }
 
-func MinInt(x,y int)int{
-	if x<y {
+func Min[T Number](x, y T) T {
+	if x < y {
 		return x
 	}
 	return y
 }
 
-func Lerp(t,a,b float64)float64{
-	return a + t * (b - a);
+func Lerp[T Number](t float64, a, b T) T {
+	return T(float64(a) + t*float64(b-a))
 }
 
-func Range(start,end float64)[]float64 {
-	var s1,e1 float64
-	if start>end {
+func Range[T Number](start, end T) []T {
+	var s1, e1 T
+	if start > end {
 		s1 = start
 		e1 = end
 	} else {
@@ -63,69 +57,39 @@ func Range(start,end float64)[]float64 {
 	if leng < 0 {
 		log.Panic("arr length must >= 0")
 	}
-	ret:=make([]float64,0)
+	ret := make([]T, 0)
 
-	for i:=s1;i<e1;i++ {
+	for i := s1; i < e1; i++ {
 		ret = append(ret, i)
 	}
 	return ret
 }
 
-func RangeInt(start,end int)[]int {
-	var s1,e1 int
-	if start>end {
-		s1 = start
-		e1 = end
-	} else {
-		s1 = end
-		e1 = start
-	}
-	leng := e1 - s1
-	if leng < 0 {
-		log.Panic("arr length must >= 0")
-	}
-	ret:=make([]int,0)
-
-	for i:=s1;i<e1;i++ {
-		ret = append(ret, i)
-	}
-	return ret
-}
-
-func Step (edge float64, x float64)float64 {
-	if x<edge {
-		return 0.0
-	}
-	return 1.0
-}
-
-func StepInt(edge int, x int)int {
-	if x<edge {
+func Step[T Number](edge T, x T) T {
+	if x < edge {
 		return 0
 	}
 	return 1
 }
 
-func Clamp (v float64, min float64, max float64)float64 {
-	return math.Min(math.Max(v, min), max)
+func Clamp[T Number](v T, min T, max T) T {
+	return Min(Max(v, min), max)
 }
 
-func ClampInt (v int, min int, max int)int {
-	return MinInt(MaxInt(v, min), max)
-}
-
-func Sign (v float64)float64 {
-	if v>0 {
-		return 1.0
+func Sign[T Signed](v T) T {
+	if v > 0 {
+		return 1
 	}
-	if v<0 {
-		return -1.0
+	if v < 0 {
+		return -1
 	}
 	return 0
 }
 
-func Fract (v float64)float64 {
-	return v - math.Floor(v)
+func Floor[T Float](v T) T {
+	return T(math.Floor(float64(v)))
 }
 
-
+func Fract[T Float](v T) T {
+	return v - Floor(v)
+}
